backend/requirements/base.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Async & WebSockets
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Celery
celery>=5.3.4
redis>=5.0.1

# Search
meilisearch>=0.28.0

# Payments
stripe>=7.0.0

# Image handling
Pillow>=10.1.0

# Utilities
python-decouple>=3.8
django-mptt>=0.15.0
